add some modules
